Find a trello card by doc id
Summary: We only want to create trello cards for project docs if the card doesn't already exist. In particular, this changeset allows us to check whether there's a card on the big/proposal boards with a description that includes a specified doc id. We might want to tweak the implementation later for perf (for example caching stuff, making fewer connections), but this suffices for now. It's hack week, after all!

Test Plan:
python testrunner.py ~/khan/webapp/third_party/frankenserver .

Yes, tests! These are a little brittle, since they depend on the product planning doc and the support initiative doc always existing on the proposal and big board (respectively). Yolo!

Reviewers: ben

Reviewed By: ben

Differential Revision: https://phabricator.khanacademy.org/D20021