Restrict how much we auto-respond
Summary:
We only want to auto-respond to the first email in a thread to new-projects, and we figure that out by checking the In-Reply-To header. If the message has the header, we ignore it. If not, we try to process it. Relatedly, we remember to tack this header on to our own auto-response.

Tangentially related, we don't auto-respond if the original message has no project doc links.

Test Plan:
I'm not sure how to test the in-reply-to locally without a lot more thorough unit testing than I'm willing to do right now. :) The local inbound mail UI doesn't let us specify headers.

Visiting localhost:8000/mail, I sent a mail to new-projects@khan-big-board.appspotmail.com that had no doc links, and I checked the logs to confirm that it was ignored.

Reviewers: ben

Reviewed By: ben

Differential Revision: https://phabricator.khanacademy.org/D20032