Add retro raccoon + "Create your retro doc" link to trello
Summary: ... and remove it when a retro doc *is* created by going through the /retro/create handler. This can happen either if the user clicks on the email link OR this trello link.

Test Plan:
Create a dummy trello card with "Monkey" in the name
With this as the description:
[![](http://khan-big-board.appspot.com/images/project-platypus.png)](https://docs.google.com/document/d/1yc5cQvu9JZefWISJcE9KQGW8SrDiTADOVNiEG-IogAs)[Project doc](https://docs.google.com/document/d/1yc5cQvu9JZefWISJcE9KQGW8SrDiTADOVNiEG-IogAs)

Add it to the completed board
https://trello.com/b/bJhNUT7y/pipeline-3-land-of-completed-project

Use that card's id and add that into retrospective_test.py

Run tests, and now see that there is a "Create your retro doc" raccoon + label!
https://s3.amazonaws.com/uploads.hipchat.com/6574/17223/BrroyLfDPd0oMUD/upload.png

Now @unittest.skip this test: test_sending_retro_reminder_for_card
And unskip this test: test_creating_retro_doc_for_card

See that the description now links to the newly minted retro doc!
https://s3.amazonaws.com/uploads.hipchat.com/6574/17223/b8maiG2Rzt8rYqi/upload.png

python testrunner.py ~/khan/webapp/third_party/frankenserver . retrospective_test.py

Reviewers: ben

Reviewed By: ben

Differential Revision: https://phabricator.khanacademy.org/D20077