Update six dependency
The https://github.com/kelp404/six.git repo has disappeared! I suppose we should have known that was going to happen since the username was kelp404... It looks like the canonical location of this libarary is https://bitbucket.org/gutworth/six, but that seems to only be available via Mercurial.

So instead of trying to use git-hg or something to track that as a submodule, I just copied the file we need, six.py, into third-party. I used the version of this file that was deployed on appengine, so it should be whatever file we were using prior to kelp404 404ing the github repo.

Test Plan:
Run `python testrunner.py ~/khan/webapp/third_party/frankenserver .`
See that the only test failures seem to be related to querying for google apps users, but I think that was already broken
 ¯\_(ツ)_/¯